[#35]; refactor: 칼로리 API 클래스 이름 및 위치 변경
